Time: 0 ms (100.00%) | Memory: 22 MB (15.89%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    bool dfs(TreeNode* root, long long low, long long high){
+    bool dfs(TreeNode* root, long low, long high){
         if(!root) return true;
         if(root->val <= low || root->val >= high) return false;
 
@@ -19,6 +19,6 @@ public:
                dfs(root->right, root->val, high);
     }
     bool isValidBST(TreeNode* root) {
-      return dfs(root, LLONG_MIN, LLONG_MAX);
+        return dfs(root, LONG_MIN, LONG_MAX);
     }
 };
